change max_score - line 243
Change max_score in line 243 to accommodate smaller negative values. For example, when using 'mean_squared_error', the call to get_scorer results in a make_scorer function with 'greater_is_better = False', such that values less than -1 can occur (the sign is flipped on the result of mse). Changed to float('-inf'), or could be -10**9, etc.