change max_score - line 243 (#89)
* change max_score - line 243

Change max_score in line 243 to accommodate smaller negative values. For example, when using 'mean_squared_error', the call to get_scorer results in a make_scorer function with 'greater_is_better = False', such that values less than -1 can occur (the sign is flipped on the result of mse). Changed to float('-inf'), or could be -10**9, etc.

* Added a unit test function to 'test_sequential_feature_selector.py', function 'test_regression_in_range', simply tests the k-range feature in the regression setting, using mean-squared-error. Previously raised an exception when using a loss function in combination with a range for k_features.

* Add additional assertions to differentiate between fixed k and k in range.Proper formatting