Updates plot_decision_regions to support >2 features
plot_decision_regions should behave exactly the same for the case when
the number of training features (X.shape[1]) is 1 or 2. This commit
adds the additional keyword arguments feature_index and
filler_feature_dict. Feature_index is an array-like object with two
items that specify the features indices in X to be plotted on the
x and y axis of the decision regions plot. Filler_feature_dict is a
dictionary of feature index-value pairs that will be filled in for the
features not on the x and y axis.