Let apriori always use low_memory processing
Thanks to previous optimizations, processing with low_memory=True is
now as efficient as with low_memory=False, and allows to process much
larger datasets.

Removing processing with low_memory=False makes code simpler.

The downside is that we do not know in advance the number of itemsets to
process, thus it is displayed afterwards.  We now display the number of
itemsets after prune step.
Note that commit 2f928cb introduced a bug, the number of processing
combinations was multiplied by itemset's length.

Since vectorized operations are no more performed on frequent itemsets,
they are stored as list of tuples.