Update decision_regions.py
found that the bottleneck for the decision_regions drawing was due to the slow prediction. to fix it, i added ray parallelization and distributed the iterative prediction on all available CPUs (local or within cluster). i have tested the changes and so far there doesn't seem to be any issues. the speed gain i measured on a small dataset was 21X the original (449 seconds vs 21 seconds on local HPC instance of 36 cores). the large dataset >2,000 input took days(>5 days) to draw, now its taking ~12 minutes on a HPC Cluster (72CPUs). 
unfortunately, due to not being a programmer, I think my implementation needs a lot of optimization and its definitely not the most pythonic way.
I failed to implement the parallelization using python multiprocessing module due to pickling issue. hence using ray algorithm.