Fix fpgrowth for sparse matrices
While this function technically supports sparse matrices, it calls
`pd.DataFrame.values` in order to get the number of rows in the
DataFrame. Unfortunately, `.values` forces the entire DataFrame to be
converted into a non-sparse 2D numpy array, which obliviates the whole
purpose behind using a sparse matrix in the first place. It's a simple
fix though.